Fix Elpian VM AST not reacting to UI interactions
Root cause: VM AST example programs rendered UI elements without `events`
properties, so taps/clicks had no handler to route to the VM. Unlike
QuickJS examples which had event bindings, VM AST counter/theme programs
relied solely on external Flutter buttons.

Changes:
- vm_example.dart: Add in-view tap events to counter (tappable Container +
  in-card Button) and theme (tappable Container) programs
- wasm_ffi.rs: Change cb_id from i64 to i32 in WASM FFI to fix
  BigInt/Number type mismatch (Dart int.toJS creates JSNumber, but
  wasm-bindgen maps i64 to BigInt, causing silent failures on web)
- executor.rs: Remove leftover debug println

+++ rust/src/api/wasm_ffi.rs
@@ -46,8 +46,8 @@ mod wasm {
     }
 
     #[wasm_bindgen]
-    pub fn elpian_wasm_execute_func(machine_id: String, func_name: String, cb_id: i64) -> String {
-        result_to_json(execute_vm_func(machine_id, func_name, cb_id))
+    pub fn elpian_wasm_execute_func(machine_id: String, func_name: String, cb_id: i32) -> String {
+        result_to_json(execute_vm_func(machine_id, func_name, cb_id as i64))
     }
 
     #[wasm_bindgen]
@@ -55,10 +55,10 @@ mod wasm {
         machine_id: String,
         func_name: String,
         input_json: String,
-        cb_id: i64,
+        cb_id: i32,
     ) -> String {
         result_to_json(execute_vm_func_with_input(
-            machine_id, func_name, input_json, cb_id,
+            machine_id, func_name, input_json, cb_id as i64,
         ))
     }
